+++ src/composables/subscribe.ts
@@ -3,6 +3,7 @@ import { RouteLocationRaw } from "vue-router"
 import { Profile, ProfilePaymentOption } from "@/api/users"
 import { useActorHandle } from "@/composables/handle"
 import { useInstanceInfo } from "@/composables/instance"
+import { useCurrentUser } from "@/composables/user"
 
 interface SubscriptionLink {
   type: "monero",
@@ -12,6 +13,7 @@ interface SubscriptionLink {
 export function useSubscribe() {
   const { getActorLocation } = useActorHandle()
   const { getBlockchainInfo } = useInstanceInfo()
+  const { currentUser } = useCurrentUser()
 
   function getSubscriptionLink(profile: Profile): SubscriptionLink | null {
     for (const option of profile.payment_options) {
@@ -20,7 +22,7 @@ export function useSubscribe() {
         option.name === "MoneroSubscription" &&
         option.href
       ) {
-        // TODO: drop support for pre-FEP-0837 links
+        // "MoneroSubscription" indicates a pre-FEP-0837 payment link
         return {
           type: "monero",
           location: option.href,
@@ -31,9 +33,16 @@ export function useSubscribe() {
           // Local subscription option, but subscription feature is disabled
           continue
         }
+        let location
+        if (currentUser.value === null && option.object_id) {
+          // Remote profile / option, viewing as a guest
+          location = option.object_id
+        } else {
+          location = getActorLocation("profile-subscription", profile)
+        }
         return {
           type: "monero",
-          location: getActorLocation("profile-subscription", profile),
+          location,
         }
       }
     }